+++ RELEASE-NOTES.md
@@ -32,6 +32,7 @@
 ## Fixes and improvements
 * Fixed problem with whitespaces in `snow connection add` command
 * Added check for the correctness of token file and private key paths when addind a connection
+* Improved error message for incompatible parameters.
 
 
 # v2.7.0